# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Recall API
  version: "1"
  description: |
    HTTP REST + SSE surface for the Recall Overwatch metrics app.

    Available in two flavors:
      * **Server binary** (`Recall-server`) — exposes this API by default
      * **Wails desktop binary with `--server`** — same API, same routes

    Conventions:
      * Every JSON endpoint is mounted under the `/api/v1/` prefix —
        the version is bumped on breaking changes.
      * Resources are nouns; sub-resources hang off the parent
        (`/matches/{match_key}/visibility`, `/settings/tesseract`).
      * Methods reflect intent: `GET` to read, `PUT` to replace/upsert,
        `DELETE` to clear or wipe, `POST` to start a server-side action
        that doesn't map to a single resource (the parse run).
      * Writes return `204 No Content` when there is no useful body
        to echo back; `202 Accepted` for actions whose effect is
        asynchronous (parse).
      * Errors come back as `application/problem+json` (RFC 9457) with
        the appropriate 4xx/5xx status. `type` is a stable URI under
        `https://github.com/sound-barrier/recall/problems/{slug}` that
        clients can switch on; body-validation problems carry a
        per-field `errors` array, and the data-update 422 carries a
        `failed_assets` list (RFC 9457 §3.2 extension members). The one
        exception is `405 Method Not Allowed`, emitted plain by the
        router.
...

...
    `RECALL_SERVER_ADDR`.
  license:
...
servers:
...
paths:
...
  /api/v1/matches/{match_key}:
    parameters:
      - $ref: "#/components/parameters/MatchKey"
    get:
      operationId: GetMatchByKey
      tags: [Matches]
      summary: Read a single match by key
      description: |
        Returns the aggregated `MatchRecord` for one match. Identical
        wire shape to a single entry from `GetMatchResults`, with the
        same read-time inference + child-table folding semantics —
        consumers that need just one match can avoid pulling the
        whole corpus.

        404 on unknown keys (the `ErrMatchNotFound` sentinel routes
        through `errors.Is` to keep honest "not in corpus" out of
        the 500 catch-all).
      responses:
        "200":
          description: Match found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Match not found.
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: HardDeleteMatch
      tags: [Matches]
      summary: Hard-delete a single match
      description: |
        Removes every trace of the match from the database: rows
        across all five parent screenshot tables (children CASCADE),
        the `match_annotations` row, and the `hidden_matches` flag.
        Source PNG files on disk are NOT touched, but `LoadAllFilenames`
        will then re-classify them as new and the next parse will
        re-ingest. Surfaced by the Hidden drawer's **Delete forever**
        affordance — the user has already moved the match to the
        archive and explicitly confirmed.

        Idempotent: unknown keys return 204.
      responses:
        "204":
          description: Match removed (or already absent).
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/parses:
    post:
      operationId: ParseScreenshots
      tags: [Pipeline]
      summary: Run a one-shot parse
      description: |
        Walks the configured screenshots folder, OCRs anything not
        already in the DB, and writes the merged rows. Idempotent —
        re-running won't duplicate matches already on record.

        `?scope=all` switches into "re-parse all": every PNG in the
        watched folder gets re-OCR'd regardless of whether it's
        already in the per-type tables. Use after a parser-tightening
        release (e.g. the hero-fuzzy-match length-gate that stopped
        Miyazaki being attributed to Mei) to retroactively correct
        older rows. The Upsert clauses are idempotent on filename so
        existing rows are rewritten in place; match annotations /
        queue / play-mode overrides / hidden flags / reviews all key
        on match_key and survive the re-parse. ~1 s per screenshot.

        Runs the parse in a BACKGROUND goroutine and returns
        `202 Accepted` **immediately** — the request is not held open
        for the OCR loop. Subscribe to `GET /api/v1/events` for progress
        + completion (`parse-progress` / `parse-complete`), and poll
        `GET /api/v1/parses/active` to resync after a reconnect or page
        reload (SSE is not replayed). Re-fetch `GET /api/v1/matches` on
        completion.

        `409 Conflict` when a parse is already in flight (single-flight,
        fail-fast) or the screenshots folder is unset/unreadable.
      parameters:
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum: [new, all]
          description: |
            `new` (default) parses only files not in the DB.
            `all` re-OCRs every file. Other values return 400.
      responses:
        "202":
          description: Parse accepted and started; subscribe to events for progress + completion.
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/settings/screenshots-folder:
    get:
      operationId: GetScreenshotsFolder
      tags: [Settings]
      summary: Get the configured screenshots folder
      description: |
        Returns the path Recall is currently reading from. Can be
        absolute (e.g. `/Users/.../Overwatch/ScreenShots/Overwatch`) or
        relative to the server's working directory.
      responses:
        "200":
          description: Current path.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderPath"
    put:
      operationId: SetScreenshotsFolder
      tags: [Settings]
      summary: Set the screenshots folder
      description: |
        Persists to `settings.json` and re-targets the file watcher if
        it's armed. Path can be absolute or relative to the server's
        working directory.

        Validation (400 on failure):
        - Only filesystem-safe characters (alphanumerics, `_ . / \ : - <space>
          ( ) + ' ,`). Shell metacharacters, NUL bytes, and control codes
          are rejected.
        - Must be in canonical form — `filepath.Clean(path) == path`
          (no redundant slashes or `..` traversal segments).
        - Must point at an existing directory.

        The echoed-back `path` is the cleaned form actually stored.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  minLength: 1
                  example: /Users/jacob/Documents/Overwatch/ScreenShots/Overwatch
      responses:
        "200":
          description: New path echoed back.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderPath"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: ResetScreenshotsFolder
      tags: [Settings]
      summary: Clear the configured screenshots folder
      description: |
        Symmetric with `DELETE /api/v1/settings/tesseract` — the user-
        set override is the thing being deleted. There's no platform
        default to fall back to; the natural empty state is "no folder
        configured" and the user re-picks via Detect / Change.

        Also tears down the file watcher if it was armed against the
        now-cleared path so we don't leak an fsnotify handle pointing
        at an orphaned dir.
      responses:
        "204":
          description: Cleared.
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/settings/tesseract:
    get:
      operationId: GetTesseractSettings
      tags: [Settings]
      summary: Get the OCR engine status
      description: |
        Reports the configured Tesseract path, whether it resolves to a
        working binary, the reported version, whether that version is
        officially supported (5.x only), the platform default path, and
        any error from the last detection attempt.

        When `found` is true but `supported` is false (e.g. a 3.x or
        4.x binary), the UI shows a caution banner and prompts for
        confirmation before running a parse — results may be incorrect
        with unsupported versions.
      responses:
        "200":
          description: Current engine status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TesseractStatus"
    put:
      operationId: SetTesseractPath
      tags: [Settings]
      summary: Set the Tesseract binary path
      description: |
        Re-runs detection on the new path. The status returned reflects
        whether the binary at that path responds to `--version`.

        Validation (400 on failure — the path is rejected before
        `exec.Command` is invoked, so a malformed path never reaches a
        syscall):
        - Only filesystem-safe characters (alphanumerics, `_ . / \ : - <space>
          ( ) + ' ,`). Shell metacharacters, NUL bytes, and control codes
          are rejected.
        - Must be in canonical form — `filepath.Clean(path) == path`.
        - Must be an absolute path.
        - Basename must be exactly `tesseract` or `tesseract.exe` — this
          pins the endpoint to "swap which Tesseract is used" and
          forbids using it to launch arbitrary binaries.

        `DELETE /api/v1/settings/tesseract` bypasses this validation
        because the platform default (`defaultTesseractPath()`) is
        trusted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  minLength: 1
                  example: /opt/homebrew/bin/tesseract
      responses:
        "200":
          description: Re-detected status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TesseractStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: ResetTesseractPath
      tags: [Settings]
      summary: Reset the Tesseract path to the platform default
      description: |
        Replaces the configured path with the platform default (Homebrew
        on macOS, `/usr/bin/tesseract` on Linux, the Windows installer
        path on Windows) and re-runs detection. Modeled as a DELETE
        because it removes the user-set override, returning the field
        to its implicit default.
      responses:
        "200":
          description: Re-detected status against the default path.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TesseractStatus"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/profiles/active:
    put:
      operationId: SwitchProfile
      tags: [Profiles]
      summary: Switch the active profile
      description: |
        Tears down the App's in-memory state tied to the previous
        profile (settings, store, watcher, metrics endpoint) and
        re-initializes everything against the new profile's directory.
        The wire response carries the new state so the caller can update
        UI without a round-trip.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: "main"
      responses:
        "200":
          description: Switched.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfilesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/profiles/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    put:
      operationId: RenameProfile
      tags: [Profiles]
      summary: Rename a profile
      description: |
        Renames the on-disk directory + updates `profiles.json`. If
        the renamed profile is the active one, the in-memory store is
        closed before the directory rename and re-opened against the
        new path so settings + DB carry through transparently. Returns
        the updated active + list.

        Idempotent when `new_name == old`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_name]
              properties:
                new_name:
                  type: string
                  description: |
                    New profile name. Same regex as creation:
                    `^[a-zA-Z0-9][a-zA-Z0-9_-]{0,39}$`.
                  example: "silentstorm"
      responses:
        "200":
          description: Renamed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfilesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: A profile with that new name already exists.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: DeleteProfile
      tags: [Profiles]
      summary: Delete a non-active profile and wipe its directory
      description: |
        Removes the profile from the list AND removes its directory
        tree (including the SQLite DB + settings) from disk. The
        active profile cannot be deleted — callers must
        `PUT /api/v1/profiles/active` to a different profile first.
      responses:
        "204":
          description: Deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Profile is active; switch away first.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    MatchKey:
      in: path
      name: match_key
      required: true
      schema:
        type: string
        minLength: 1
      description: |
        Match identity — same `match_key` value exposed in
        `MatchRecord`. URL-safe: the canonical form replaces every
        legacy colon separator with a dash, so no percent-encoding
        is required for paste-in-URL use
        (e.g. `match-2026-05-10T22-21-11`). For `unmatched-<filename>`
        and `ambiguous-<filename>` variants the embedded filename
        still needs the usual encoding for spaces / unicode.
      example: "match-2026-05-10T22-21-11"
  schemas:
...
    FolderPath:
      type: object
      required: [path]
      properties:
        path:
          type: string
          example: /Users/jacob/Documents/Overwatch/ScreenShots/Overwatch
          description: |
            The configured screenshots folder. The GET response can
            return an empty string when no folder has been configured;
            the PUT request body inlines its own minLength: 1 schema.
...
    ProfilesResponse:
      type: object
      required: [active, profiles]
      properties:
        active:
          type: string
          description: The currently-active profile name.
          example: "main"
        profiles:
          type: array
          items:
            type: string
          description: |
            Every known profile name, sorted alphabetically (the
            order matches `slices.Sort` on the slice the
            `*Profiles.List` method returns).
          example: ["alt", "main"]
...
  responses:
    BadRequest:
      description: Malformed request body or query parameters.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    Conflict:
      description: |
        The request was syntactically valid, but the resource state or
        a payload value prevents the action (e.g. duplicate profile
        name, screenshots directory not configured, invalid Tesseract
        binary path, non-candidate resolution target).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    InternalError:
      description: Unhandled server-side error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...